requirements.txt
huggingface_hub
datasets
python-dotenv
requests
